lib/string/strcpy/, lib/, src/, tests/: Move all copying APIs to a subdirectory
