+++ lib/string/strcpy/strtcpy.c
@@ -1,7 +1,5 @@
-/*
- * SPDX-FileCopyrightText: 2022-2023, Alejandro Colomar <alx@kernel.org>
- * SPDX-License-Identifier: BSD-3-Clause
- */
+// SPDX-FileCopyrightText: 2022-2024, Alejandro Colomar <alx@kernel.org>
+// SPDX-License-Identifier: BSD-3-Clause
 
 
 #include <config.h>
@@ -9,9 +7,7 @@
 #include <stddef.h>
 #include <sys/types.h>
 
-#ident "$Id$"
-
-#include "string/strtcpy.h"
+#include "string/strcpy/strtcpy.h"
 
 
 extern inline ssize_t strtcpy(char *restrict dst, const char *restrict src,
