+++ lib/string/strcpy/strtcpy.h
@@ -1,11 +1,9 @@
-/*
- * SPDX-FileCopyrightText: 2023, Alejandro Colomar <alx@kernel.org>
- * SPDX-License-Identifier: BSD-3-Clause
- */
+// SPDX-FileCopyrightText: 2023-2024, Alejandro Colomar <alx@kernel.org>
+// SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRTCPY_H_
-#define SHADOW_INCLUDE_LIB_STRTCPY_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRCPY_STRTCPY_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRCPY_STRTCPY_H_
 
 
 #include <config.h>
@@ -16,7 +14,6 @@
 #include <sys/types.h>
 
 #include "attr.h"
-#include "defines.h"
 #include "sizeof.h"
