+++ lib/string/strcpy/zustr2stp.h
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 
-#ifndef SHADOW_INCLUDE_LIB_STRING_ZUSTR2STP_H_
-#define SHADOW_INCLUDE_LIB_STRING_ZUSTR2STP_H_
+#ifndef SHADOW_INCLUDE_LIB_STRING_STRCPY_ZUSTR2STP_H_
+#define SHADOW_INCLUDE_LIB_STRING_STRCPY_ZUSTR2STP_H_
 
 
 #include <config.h>
