+++ src/gpasswd.c
@@ -36,7 +36,7 @@
 #include "exitcodes.h"
 #include "shadowlog.h"
 #include "string/sprintf/snprintf.h"
-#include "string/strtcpy.h"
+#include "string/strcpy/strtcpy.h"
 
 
 /*
