+++ src/login.c
@@ -39,8 +39,8 @@
 #include "exitcodes.h"
 #include "shadowlog.h"
 #include "string/sprintf/snprintf.h"
+#include "string/strcpy/strtcpy.h"
 #include "string/strftime.h"
-#include "string/strtcpy.h"
 
 
 #ifdef USE_PAM
