+++ src/su.c
@@ -60,7 +60,7 @@
 #include "exitcodes.h"
 #include "shadowlog.h"
 #include "string/sprintf/snprintf.h"
-#include "string/strtcpy.h"
+#include "string/strcpy/strtcpy.h"
 
 
 /*
